+++ src/hooks/postrun/notify.ts
@@ -41,9 +41,10 @@ export const hook = async (options: any) => {
     const projectName = process.env.CI_PROJECT_NAME || (await getGitRepoName()) || "Missing CI_PROJECT_NAME variable";
     const branchName = process.env.CI_COMMIT_REF_NAME || (await getCurrentGitBranch({ formatted: true })) || "Missing CI_COMMIT_REF_NAME variable";
     const envName = projectName + "/" + branchName;
+    const diffFilesTxt = diffFiles.map((diffFile) => `• ${diffFile.path.replace("force-app/main/default/", "")} (${diffFile.index})`).join("\n")
     await sendNotification({
       title: `Updates detected in org ${envName}`,
-      text: `<pre>${diffFiles.join("\n")}</pre>`,
+      text: `<pre>${diffFilesTxt}</pre>`,
       summary: `Changes on metadatas has been detected on ${envName}. You may want to have a look !`,
       buttons: [{ title: "View commit", url: jobUrl }],
       severity: "info",