+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [4.17.1] 2023-11-28
+
+- Generate CSV output for hardis:org:monitor:backup
+- Refactor git detection of created/updated/deleted files
+
 ## [4.17.0] 2023-11-28
 
 - hardis:org:backup: Monitor installed packages